requirements.txt
requests==2.31.0
python-dotenv==1.0.0
langgraph>=0.2.0
langchain>=0.2.0
langchain-openai>=0.1.0
langchain-community>=0.2.0
streamlit>=1.28.0
fpdf>=1.7.2
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
